+++ test/helper/weber_helper_resource_test.exs
@@ -10,7 +10,7 @@ defmodule WeberHelperResourceTest do
 
   test "Create script tag with default value" do
     script = script()
-    assert(script == "<script type=\"text/javascript\" src=\"/public/application.js\"></script>")
+    assert(script == "<script type=\"text/javascript\" src=\"/public/js/application.js\"></script>")
   end
 
   test "Create style tag" do
@@ -20,7 +20,7 @@ defmodule WeberHelperResourceTest do
 
   test "Create style tag with default value" do
     style = style()
-    assert(style == "<link href=\"/public/application.css\" rel=\"stylesheet\" media=\"screen\">")
+    assert(style == "<link href=\"/public/css/application.css\" rel=\"stylesheet\" media=\"screen\">")
   end
 
   test "Create favicon tag" do
@@ -41,6 +41,22 @@ defmodule WeberHelperResourceTest do
   test "Create image tag with no html_options" do
     image = image("/public/img/example.jpg")
     assert(image == "<img src=\"/public/img/example.jpg\">")
+  end
+
+  test "Create simple video tag" do
+    video = video("/public/videos/video1")
+    assert(video == "<video src=\"/public/videos/video1\"></video>")
+
+    video2 = video("/public/videos/video1.ogg")
+    assert(video2 == "<video src=\"/public/videos/video1.ogg\"></video>")
+
+    video3 = video("/public/videos/video1.ogg", [controls: "controls"])
+    assert(video3 == "<video src=\"/public/videos/video1.ogg\" controls=\"controls\"></video>")
+  end
+
+  test "Create video/source tag" do
+    video = video(["/public/videos/video1", "/public/videos/video1"])
+    assert(video == "<video><source src=\"/public/videos/video1\"><source src=\"/public/videos/video1\"></video>")
   end
 
 end